ch4/ofi: move dynamic process code into dynproc.c
Currently the code managing dynamic process is spread between ofi_init.c
and ofi_spawn.c. Move them to its own source file for better
organization.
